requirements.txt
# Requirements for running experiments
numpy
scipy
scikit-learn
matplotlib
pyyaml
torch
